+++ src/components/TabsLayout/types.ts
@@ -1,10 +1,15 @@
 import { ReactNode } from "react";
+import { ButtonPropsType } from "@sito/dashboard";
+
+export type TabButtonPropsType = Omit<ButtonPropsType, "children" | "onClick" | "type">;
 
 export type TabsLayoutPropsType = {
   tabs: TabsType[];
   defaultTab?: number;
   className?: string;
   tabsContainerClassName?: string;
+  useLinks?: boolean;
+  tabButtonProps?: TabButtonPropsType;
 };
 
 export type TabsType = {
@@ -21,4 +26,6 @@ export type TabPropsType = {
   active: boolean;
   onClick: () => void;
   siblings?: boolean;
+  useLinks?: boolean;
+  tabButtonProps?: TabButtonPropsType;
 };