+++ src/lib/entities/auth/AuthDto.ts
@@ -1,3 +1,7 @@
 import { BaseAuthDto } from "./BaseAuthDto";
 
-export type AuthDto<TExtra = Record<string, never>> = BaseAuthDto<TExtra>;
+export type AuthDto<TExtra extends object = object> = BaseAuthDto<
+  TExtra & {
+    rememberMe?: boolean;
+  }
+>;