+++ src/providers/DrawerMenuProvider.tsx
@@ -12,7 +12,16 @@ import { DrawerMenuContextType, DrawerMenuProviderPropTypes } from "./types";
 // lib
 import { SubMenuItemType } from "lib";
 
-const DrawerMenuContext = createContext({});
+const defaultDrawerMenuContext: DrawerMenuContextType<string> = {
+  addChildItem: () => {},
+  removeChildItem: () => {},
+  clearDynamicItems: () => {},
+  dynamicItems: {},
+};
+
+const DrawerMenuContext = createContext<DrawerMenuContextType<string>>(
+  defaultDrawerMenuContext
+);
 
 /**
  * Drawer Menu Provider
@@ -66,7 +75,9 @@ const DrawerMenuProvider = <MenuKeys extends string>(
   );
 
   return (
-    <DrawerMenuContext.Provider value={value}>
+    <DrawerMenuContext.Provider
+      value={value as unknown as DrawerMenuContextType<string>}
+    >
       {children}
     </DrawerMenuContext.Provider>
   );
@@ -77,9 +88,8 @@ const DrawerMenuProvider = <MenuKeys extends string>(
  * @returns Provider
  */
 const useDrawerMenu = <MenuKeys extends string>() => {
-  return useContext(
-    DrawerMenuContext
-  ) as unknown as DrawerMenuContextType<MenuKeys>;
+  const context = useContext(DrawerMenuContext);
+  return context as DrawerMenuContextType<MenuKeys>;
 };
 
 // eslint-disable-next-line react-refresh/only-export-components